move extras_providers module into package
